requirements.txt
spotipy
python-dotenv
requests
flask